Remove unnecessary (nuint)/(nint) casts for uint/ulong/int/long params
The generator was incorrectly casting uint/ulong to (nuint) and int/long to
(nint) when passing method parameters to MsgSend. These are fixed-size types
(uint32_t, uint64_t, etc.) that should not be converted to platform-dependent
native types.

- Remove the bool, uint/ulong→nuint, int/long→nint branches in
  CSharpEmitter.cs; all now fall through to pass values directly
- Add MsgSend overloads in ObjectiveC.cs for ulong parameters:
  MsgSend(nint, Selector, nint, ulong),
  MsgSend(nint, Selector, nint, ulong, MTLSharedEventNotificationBlock),
  MsgSendBool(nint, Selector, ulong, ulong)
- Regenerated 6 affected files

+++ Metal.NET/Metal/MTL4CommandQueue.cs
@@ -98,7 +98,7 @@ public class MTL4CommandQueue(nint nativePtr, NativeObjectOwnership ownership) :
 
     public void SignalEvent(MTLEvent @event, ulong value)
     {
-        ObjectiveC.MsgSend(NativePtr, MTL4CommandQueueBindings.SignalEvent, @event.NativePtr, (nuint)value);
+        ObjectiveC.MsgSend(NativePtr, MTL4CommandQueueBindings.SignalEvent, @event.NativePtr, value);
     }
 
     public unsafe void UpdateBufferMappings(MTLBuffer buffer, MTLHeap heap, MTL4UpdateSparseBufferMappingOperation[] operations)
@@ -119,7 +119,7 @@ public class MTL4CommandQueue(nint nativePtr, NativeObjectOwnership ownership) :
 
     public void Wait(MTLEvent @event, ulong value)
     {
-        ObjectiveC.MsgSend(NativePtr, MTL4CommandQueueBindings.Wait, @event.NativePtr, (nuint)value);
+        ObjectiveC.MsgSend(NativePtr, MTL4CommandQueueBindings.Wait, @event.NativePtr, value);
     }
 
     public void Wait(MTLDrawable drawable)
